+++ package.json
@@ -15,7 +15,6 @@
     "bootstrap-4-grid": "^3.1.0",
     "case-sensitive-paths-webpack-plugin": "2.1.1",
     "chalk": "1.1.3",
-    "chrome-extension-async": "^3.3.2",
     "css-loader": "0.28.7",
     "dotenv": "4.0.0",
     "dotenv-expand": "4.2.0",
