+++ public/newtab.html
@@ -14,7 +14,6 @@
     <noscript>
       You need to enable JavaScript to run this app.
     </noscript>
-    <script src="%PUBLIC_URL%/app/chrome-extension-async.js"></script>
     <script src="%PUBLIC_URL%/app/d3.min.js"></script>
     <div id="root"></div>
